+++ src/App/Send/SendForm.js
@@ -34,7 +34,7 @@ import {
 
 const SendFormComponent = styled.form({
   maxWidth: 800,
-  margin: '0 auto',
+  margin: '-.5em auto 0',
 });
 
 const SendFormButtons = styled.div({