+++ src/shared/utils/etc.js
@@ -28,12 +28,41 @@ export function passRef(el, ref) {
 }
 
 export function debounced(fn, ms) {
-  let timeout;
-
+  let timerId;
   return function() {
     const functionCall = () => fn.apply(this, arguments);
-
-    clearTimeout(timeout);
-    timeout = setTimeout(functionCall, ms);
+    clearTimeout(timerId);
+    timerId = setTimeout(functionCall, ms);
+    return timerId;
   };
+}
+
+export function throttled(fn, ms) {
+  let lastTimerId;
+  let lastRan;
+  return function() {
+    const context = this;
+    const args = arguments;
+    if (!lastRan) {
+      fn.apply(context, args);
+      lastRan = Date.now();
+    } else {
+      clearTimeout(lastTimerId);
+      lastTimerId = setTimeout(function() {
+        fn.apply(context, args);
+        lastRan = Date.now();
+      }, ms - (Date.now() - lastRan));
+    }
+    return lastTimerId;
+  };
+}
+
+export function limitDecimal(num, dec) {
+  if (typeof num === 'number' && typeof dec === 'number') {
+    const [int, decimal] = String(num).split('.');
+    if (decimal && decimal.length > dec) {
+      return num.toFixed(dec);
+    }
+  }
+  return num;
 }